Refactor dto package
