+++ src/main/java/com/bh/rms/controller/PurchaseController.java
@@ -2,8 +2,8 @@ package com.bh.rms.controller;
 
 import com.bh.rms.domain.aggregate.purchase.Purchase;
 import com.bh.rms.domain.aggregate.purchase.service.PurchaseService;
-import com.bh.rms.domain.aggregate.purchase.service.dto.PurchaseCreateRequest;
-import com.bh.rms.domain.aggregate.purchase.service.dto.PurchaseUpdateRequest;
+import com.bh.rms.web.dto.PurchaseCreateRequest;
+import com.bh.rms.web.dto.PurchaseUpdateRequest;
 import jakarta.validation.Valid;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
