+++ src/test/java/com/bh/rms/domain/aggregate/purchase/service/PurchaseServiceTest.java
@@ -1,9 +1,8 @@
 package com.bh.rms.domain.aggregate.purchase.service;
 
-import com.bh.rms.domain.aggregate.purchase.Purchase;
 import com.bh.rms.domain.aggregate.purchase.exception.PurchaseNotFoundException;
 import com.bh.rms.domain.aggregate.purchase.infra.PurchaseRepository;
-import com.bh.rms.domain.aggregate.purchase.service.dto.PurchaseCreateRequest;
+import com.bh.rms.web.dto.PurchaseCreateRequest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -16,7 +15,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
